Moved list of infoarena controllers from MyTextile.php to config.php as suggested in the code review. Sadly, php's define() doesn't work with arrays
Changed how index.php handles links a bis. It first checks if $url_root is in $IA_CONTROLLERS and assumes url is a simple textblock if not.
When adding new functionality to the site developers will have to add controller to $IA_DIRECT_CONTROLLERS or $IA_CONTROLLERS.

Removed 'userinfo' controller from list since it doesn't exist :)

Changed execute_view to execute_view_die in 4 controllers, since all the others are this way.



git-svn-id: https://hackers.infoarena.ro/dav-svn/trunk@915 ddd5704e-cc05-0410-ac10-d31acb56fd8e