Fix SMF not updating it's member stats.
Currently, SMF reports it has 7000 members and the latest member registered in February, 2008.

Apparently, SMF only updates it's member statistics when a new user is registered, so I adapted some of it's update statistics code and made it run when a new user is registered.

Reviewed: http://reviewboard.infoarena.ro/r/91/



git-svn-id: https://hackers.infoarena.ro/dav-svn/trunk@1058 ddd5704e-cc05-0410-ac10-d31acb56fd8e