Made caching functions more resistent to failures.
Whenever memcache_set would fail (with a timeout for example), it would
raise a PHP warning which would cause termination of the running script.
This is why the evaluator has been crashing daily for the past few
weeks. Silenced all cache functions with a @ since they already return
TRUE or FALSE on success.

Changes were made to config.php.sample so that task and round cache key
TTLs are larger when using memcached, since the evaluator is able to
invalidate changed data.

Diff contains a fix from live for determining if the website is running
under https when behind a reverse proxy (such as nginx).

Most of the diff's changes are lint fixes.

REVIEW URL: http://reviewboard.infoarena.ro/r/201/

git-svn-id: https://hackers.infoarena.ro/dav-svn/trunk@1178 ddd5704e-cc05-0410-ac10-d31acb56fd8e