Support for separate domains for http and https traffic.
Summary:
Restrict certain endpoints to https and others to http. Whenever a
user goes on an endpoint with the wrong protocol or wrong host name,
they get redirected.

This will allow us to enable cloudflare on infoarena.ro. We'll keep
http traffic on www.infoarena.ro and https traffic on infoarena.ro
and use a cookie domain of .infoarena.ro

Test Plan: Deployed on live.

Reviewers: adrian.budau

Reviewed By: adrian.budau

CC: echipa-infoarena

Differential Revision: https://hackers.infoarena.ro/D17