Filter virtual contests from user stats page.
Summary:
We should only display classic and penalty rounds on a user's stats page. The
feature was requested by a couple of users here:
http://www.infoarena.ro/blog/imbunatatim-infoarena.

Not sure how the round_id can become NULL, so I removed that check as well. Let
me know if I'm missing anything.

Test Plan:
I added an entry for my user_id and a virtual contest round in
ia_score_user_round_task and checked that it doesn't show up on
/utilizator/pauldb?action=stats.

Reviewers: bogdan2412, adrian.budau

Reviewed By: bogdan2412

CC: echipa-infoarena

Differential Revision: https://hackers.infoarena.ro/D18