removed round tags since it was breaking the parameter submit for round edit
Summary:
round-edit page, params tab: the dual select input was calling the parent form's onSubmit method. In this method there was a piece of code that checks if autocomplete is still running for round tags, an input we do not even use.
Occasionally that script seems to report a false positive even though autocomplete should not be running.
Removed that piece of code as well as the tags input field for round-edit page.

Test Plan: manual testing in browser

Reviewers: bogdan2412, adrian.budau

Reviewed By: adrian.budau

Subscribers: echipa-infoarena

Differential Revision: https://hackers.infoarena.ro/D32