Added problem statistics feature
Summary:
This diff roughly contains: a script to update the database for our needs;
controllers and views for the new page; code that updates the statistics in
the eval.
We create, update and query the table ia_score_task_top_users because it is
a lot faster than recomputing the statistics from the ia_job and ia_job_test
tables.
A lot of the changes are just code formatting changes because the linter complained.
Sorry for the huge diff :-)

Test Plan:
You need to run scripts/compute-statistics (maybe not the best name)
Then access /statistici_problema?task={TASK_ID}.

Reviewers: adrian.budau, vduta

Reviewed By: vduta

Subscribers: adrian.budau, echipa-infoarena

Differential Revision: https://hackers.infoarena.ro/D33