Problem statistics extension
Summary:
New features added:
- the statistics should work for every archive and every contest
- every entry in the rankings contains a link towards the coresponding job

On a closer inspection, the queries for computing the rankings were wrong (submit_time was not correctly calculated).
I had to adjust them anyway for the new features, so I decided to rethink everything to make the code cleaner.
Instead of the SQL queries doing the job, now the queries are simple and php does the processing.

So you will have to rerun the compute-statistics script, but without the last heavy query.

It is hard to test everything so please review the code carefully :-)

Test Plan: Run scripts/compute-statistics and then you can access /statistici_problema?task={TASK_ID}&round={ROUND_ID}

Reviewers: adrian.budau, vduta

Reviewed By: vduta

Subscribers: bogdan2412, tiberiu, echipa-infoarena-list

Differential Revision: https://hackers.infoarena.ro/D36